+++ Mods ChoGGi/Canadian Space Agency 2/items.lua
@@ -195,4 +195,8 @@ return {
 		}),
 	}),
 
+	PlaceObj("ModItemCode", {
+		"FileName", "Code/Script.lua",
+	}),
+
 }