+++ Mods ChoGGi/Fix Unrepairable Attack Rovers/metadata.lua
@@ -1,0 +1,24 @@
+return PlaceObj("ModDef", {
+	"title", "Fix Unrepairable Attack Rovers",
+	"id", "ChoGGi_FixUnrepairableAttackRovers",
+	"steam_id", "2967276566",
+	"pops_any_uuid", "ff812298-91d8-406f-8082-3e8431096b6c",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.png",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagOther", true,
+	"description", [[
+Allows Attack Rovers to be repaired when you can't repair them.
+I don't want to say too much as they're from one of the mysteries.
+
+
+https://www.reddit.com/r/SurvivingMars/comments/12vkbrb/
+]],
+})