+++ Mods ChoGGi/Partial Terraforming/metadata.lua
@@ -1,0 +1,24 @@
+return PlaceObj("ModDef", {
+	"title", "Partial Terraforming",
+	"id", "ChoGGi_PartialTerraforming",
+	"steam_id", "2972432670",
+	"pops_any_uuid", "0a41cb47-8c5d-4e5e-82ea-af7d129f9072",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagTerraforming", true,
+	"description", [[
+Ancient terraforming device = all stats go to 40% (set to 40, not add 40).
+Stop Blue skies, Cold waves/Toxic rains never stop.
+
+
+Requested by Paladin Boyd.
+]],
+})