+++ Mods ChoGGi/Trade Pad Cheats/metadata.lua
@@ -1,0 +1,26 @@
+return PlaceObj("ModDef", {
+	"title", "Trade Pad Cheats",
+	"id", "ChoGGi_TradePadCheats",
+	"steam_id", "2963471158",
+	"pops_any_uuid", "cb53c375-6e71-4028-9370-8632d021ff80",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+Cheat(s) for Trade Pads. See Mod options to cheat.
+
+Mod Options:
+Min AI Export Amount: Rivals will not trade a resource if they have under this amount of said resource.
+
+
+Not yet requested by Far_Communication372.
+]],
+})